Defect 189: Vlan status is up ...
Vlan status is up with no ports assigned

Diagnosed as an artifact of adding the bridge_normal
port and interface. The port is all default values,
so it is by default in trunk mode and has no trunk
vlans defined, so it is in *all* vlans, resulting in
an incorrect status. Solution is to ignore the
port with the name "bridge_normal" (there's no other
way to identify the port).

Change-Id: Ifcf0ccb961e2341e00ff9c62ec892ea28dabb23b
Signed-off-by: Dan Cripe <dan.cripe@hpe.com>