Internal VLAN range cannot be changed if interface vlan is configured
Tags: fix, dev
TG-916. Internal VLAN allocation issue between interface VLAN
and physical L3 interface. If a interface VLAN is created then
that VLAN shouldnt be used as an internal VLAN for a physical
l3 interface. But range of internal VLAN can be changed
irrespective of interface VLAN being part of the range

Change-Id: Ia18ae604e5aee5c68d520ed7f3600240133e2abb
Signed-off-by: Sandeep Nagapattinam <sandeep.nagapattinam@hpe.com>